Nail down when Bools are i1 or i8
Are `bool`s i1 or i8 in LLVM? Some investigation reveals that it depends!
Moreover, `copilot-verifier` almost gets things right, but not quite. In this
commit, I:

* Write up my findings in `Note [How LLVM represents bool]`, and
* Tweak the implementations of `copilotTypeToMemType` and
  `copilotTypeToLLVMType` to adhere to that Note. In the process, I renamed the
  clunkily named `copilotTypeToMemType'` to `copilotTypeToMemTypeBool8` to
  better reflect its intended purpose. I also created a
  `copilotTypeToLLVMTypeBool8` function, which works analogously.

This doesn't fix any bugs in `copilot-verifier`'s current feature set, but it
will become important in a subsequent commit that adds support for arrays, as
arrays of `bool`s in LLVM become arrays of `i8`, not `i1`.