Support structs
This relies on a somewhat non-trivial codegen change on the `copilot` side
that makes struct-typed arguments in trigger functions passed by reference,
not by value. Doing so makes it much more predictable what shape the arguments
will take on in LLVM (see `Note [Arrays and structs]`, which was previously
`Note [Arrays]`). Once this change is in place, supporting arrays proves quite
straightforward, as it mirrors the codepaths for arrays quite closely.

Fixes #9.