Control verbosity more in logging
This adds a new `Verbosity` data type and a corresponding `verifyWithVerbosity`
function that controls whether `copilot-verifier` should produce logs or not.
We want to make the `Verbosity` be `Quiet` primarily in the test suite so that
the diagnostic output does not drown out the output from `tasty`.

This is a first step towards the first bullet point in #13.