Add option for assuming side conditions from partial operations
As described in a new section of the `README`, there are two general approaches
for handling partial operations. This behavior can now be toggled using the new
`assumePartialSideConds` option in the `VerifierOptions`. The default option is
to error out whenever any partial operation is encounted on the C side. Another
setting is to assume that any invocation of a partial operation in the Copilot
spec will always be on well defined inputs. That way, partiality on the C side
will be OK as long as it matches up with a corresponding partial function on
the Copilot side (i.e., the programs are "crash-equivalent").

Fixes #20.