Ensure trigger functions are called exactly once
To do so, we track how many times each trigger function is invoked with a
`Nat`-valued global variable rather than a simple `Bool`-valued global. At
the end of simulation, we ensure the value in the global is equal to
`if guard_cond then 1 else 0`. See the new
`Note [Global variables for trigger functions]`.

Fixes #29.