Fix minor buglet when calling verifyStepBisimulation
Previously, we were passing `emptyMem` when calling `verifyStepBisimulation`,
but we should have been passing `initialMem` instead. Using the latter is
important because it initializes memory for global variables, which is
necessary if the generated LLVM bitcode references a global variable.

This bug didn't show up in any of the repo's current examples, but it did show
up when scaling up to a more complicated example involving constant array
initializers.