Support logging raw SMT solver interactions
This adds a `logSmtInteractions` flag to `VerifierOptions` that, when enabled,
will dump each SMT solver interaction into its own file under `results/`. See
the Haddocks for `logSmtInteractions` for the full details. Implementing this
requires bumping the `crucible` submodule to bring in the changes from
GaloisInc/crucible#1067, which adds support for logging offline SMT solver
interactions.

This adds some information that will be needed for #23.