Pass -ffp-contract=off to Clang when compiling C code
Clang 14+ enable `-ffp-contract=on` by default, which cause floating-point
expressions such as `a + b * c` to be optimized to fused multiply-add LLVM
intrinsics (e.g., `llvm.fmuladd.*`). This is a problem for `copilot-verifier`,
which critically relies on preserving the order of additions and
mulitplications. Thankfully, we can override this option by explicitly passing
`-ffp-contract=off` to Clang.