Better user feedback upon a successful verification
Upon a successful run of the verifier, we now print out a general description
of what `copilot-verifier` did, along with brief explanations of what the
various proof goals indicate. This should make it easier to interpret the
results if you are not an expert in formal verification.

We also print out a suggestion to turn the verbosity to `Noisy` if you want a
more detailed look into what each proof goal means.