Better distinguish between durable and non-durable proof goals. Refs #51.
Previously, the verifier was inconsistent about whether it would log certain
types of proof goals. As discussed in #51, there are really two types of proof
goals:

1. Goals that the verifier asserts directly as part of the correspondence proof.
2. Goals that arise during symbolic execution of the `step()` function to ensure
   memory safety.

We always want to log goals of type (1), as they are core to the correspondence
proof that the verifier is computing. Due to the sheer volume of goals of type
(2), however, we only want to log type-(2) goals if they are non-trivial (i.e.,
not just the `true` predicate in `what4`).

This patch:

* Ensures that all type-(1) goals are asserting using Crucible's _durable_
  assertions so that they will always be logged, even if they are trivial.
* Makes the verifier's logs distinguish between the two types of goals when the
  `Noisy` verbosity level is chosen.