Reject multiple triggers with the same name. Refs #74.
Although `copilot-c99` added the ability to invoke multiple triggers with the
same name in https://github.com/Copilot-Language/copilot/pull/572, it is not
yet clear how best to support this in `copilot-verifier`.

In the meantime, this adds an explicit check that rules out specifications that
use multiple triggers with the same name to prevent the verifier from becoming
confused by them. The remaining task of fully supporting such specifications is
tracked in https://github.com/Copilot-Language/copilot-verifier/issues/74.