Add smtFloatMode option to VerifierOptions. Refs #79.
This allows users to switch between `FloatUninterpreted` (the default setting,
where all floating-point operations are treated as uninterpreted functions) and
`FloatIEEE` (where floating-point values are treated as IEEE-754 floats). While
`FloatIEEE` is unable to handle floating-point operations that are not native
to SMT-LIB (`sin`, `cos`, `tan`, etc.), it can handle more specifications in
which a C compiler optimizes the floating-point expressions found in the
generated C code. An example of where this is crucible can be found in the
newly added `FPNegation` example, which was inspired by Ogma.